pyproject.toml
[project]
name = "cloneme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "google-genai>=1.22.0",
    "openai>=1.93.0",
    "python-dotenv",
    "python-telegram-bot>=22.1",
]
